added trigger for rule deletion
